+++ src/Polynomials/QuinticPolynomial.h
@@ -6,12 +6,12 @@
 class QuinticPolynomial {
 public:
     QuinticPolynomial() = default;
-    QuinticPolynomial(double xs, double vxs, double axs, double xe,
-                      double vxe, double axe, double t);
-    fixp_d calc_point(double t);
-    fixp_d_d calc_first_derivative(double t);
-    fixp_d_dd calc_second_derivative(double t);
-    fixp_d_ddd calc_third_derivative(double t);
+    QuinticPolynomial(float xs, float vxs, float axs, float xe,
+                      float vxe, float axe, float t);
+    fixp_d calc_point(float t);
+    fixp_d_d calc_first_derivative(float t);
+    fixp_d_dd calc_second_derivative(float t);
+    fixp_d_ddd calc_third_derivative(float t);
 private:
     int_4_11 a0, a1, a2, a3, a4, a5;
 };
